[AIRFLOW-5919] Group tests for the Users/Roles/Perms commands
