+++ tests/core.py
@@ -18,7 +18,6 @@
 # under the License.
 
 import io
-import json
 import multiprocessing
 import os
 import pickle  # type: ignore
@@ -28,7 +27,6 @@ import subprocess
 import tempfile
 import unittest
 from datetime import timedelta
-from tempfile import NamedTemporaryFile
 from time import sleep
 from unittest import mock
 
@@ -1031,283 +1029,6 @@ class TestCli(unittest.TestCase):
         session.query(Variable).delete()
         session.commit()
         session.close()
-
-    def test_cli_create_user_random_password(self):
-        args = self.parser.parse_args([
-            'users', 'create', '--username', 'test1', '--lastname', 'doe',
-            '--firstname', 'jon',
-            '--email', 'jdoe@foo.com', '--role', 'Viewer', '--use_random_password'
-        ])
-        cli.users_create(args)
-
-    def test_cli_create_user_supplied_password(self):
-        args = self.parser.parse_args([
-            'users', 'create', '--username', 'test2', '--lastname', 'doe',
-            '--firstname', 'jon',
-            '--email', 'jdoe@apache.org', '--role', 'Viewer', '--password', 'test'
-        ])
-        cli.users_create(args)
-
-    def test_cli_delete_user(self):
-        args = self.parser.parse_args([
-            'users', 'create', '--username', 'test3', '--lastname', 'doe',
-            '--firstname', 'jon',
-            '--email', 'jdoe@example.com', '--role', 'Viewer', '--use_random_password'
-        ])
-        cli.users_create(args)
-        args = self.parser.parse_args([
-            'users', 'delete', '--username', 'test3',
-        ])
-        cli.users_delete(args)
-
-    def test_cli_list_users(self):
-        for i in range(0, 3):
-            args = self.parser.parse_args([
-                'users', 'create', '--username', 'user{}'.format(i), '--lastname',
-                'doe', '--firstname', 'jon',
-                '--email', 'jdoe+{}@gmail.com'.format(i), '--role', 'Viewer',
-                '--use_random_password'
-            ])
-            cli.users_create(args)
-        with mock.patch('sys.stdout', new_callable=io.StringIO) as mock_stdout:
-            cli.users_list(self.parser.parse_args(['users', 'list']))
-            stdout = mock_stdout.getvalue()
-        for i in range(0, 3):
-            self.assertIn('user{}'.format(i), stdout)
-
-    def test_cli_list_users_with_args(self):
-        cli.users_list(self.parser.parse_args(['users', 'list',
-                                               '--output', 'tsv']))
-
-    def test_cli_import_users(self):
-        def assertUserInRoles(email, roles):
-            for role in roles:
-                self.assertTrue(self._does_user_belong_to_role(email, role))
-
-        def assertUserNotInRoles(email, roles):
-            for role in roles:
-                self.assertFalse(self._does_user_belong_to_role(email, role))
-
-        assertUserNotInRoles(self.TEST_USER1_EMAIL, ['Admin', 'Op'])
-        assertUserNotInRoles(self.TEST_USER2_EMAIL, ['Public'])
-        users = [
-            {
-                "username": "imported_user1", "lastname": "doe1",
-                "firstname": "jon", "email": self.TEST_USER1_EMAIL,
-                "roles": ["Admin", "Op"]
-            },
-            {
-                "username": "imported_user2", "lastname": "doe2",
-                "firstname": "jon", "email": self.TEST_USER2_EMAIL,
-                "roles": ["Public"]
-            }
-        ]
-        self._import_users_from_file(users)
-
-        assertUserInRoles(self.TEST_USER1_EMAIL, ['Admin', 'Op'])
-        assertUserInRoles(self.TEST_USER2_EMAIL, ['Public'])
-
-        users = [
-            {
-                "username": "imported_user1", "lastname": "doe1",
-                "firstname": "jon", "email": self.TEST_USER1_EMAIL,
-                "roles": ["Public"]
-            },
-            {
-                "username": "imported_user2", "lastname": "doe2",
-                "firstname": "jon", "email": self.TEST_USER2_EMAIL,
-                "roles": ["Admin"]
-            }
-        ]
-        self._import_users_from_file(users)
-
-        assertUserNotInRoles(self.TEST_USER1_EMAIL, ['Admin', 'Op'])
-        assertUserInRoles(self.TEST_USER1_EMAIL, ['Public'])
-        assertUserNotInRoles(self.TEST_USER2_EMAIL, ['Public'])
-        assertUserInRoles(self.TEST_USER2_EMAIL, ['Admin'])
-
-    def test_cli_export_users(self):
-        user1 = {"username": "imported_user1", "lastname": "doe1",
-                 "firstname": "jon", "email": self.TEST_USER1_EMAIL,
-                 "roles": ["Public"]}
-        user2 = {"username": "imported_user2", "lastname": "doe2",
-                 "firstname": "jon", "email": self.TEST_USER2_EMAIL,
-                 "roles": ["Admin"]}
-        self._import_users_from_file([user1, user2])
-
-        users_filename = self._export_users_to_file()
-        with open(users_filename, mode='r') as file:
-            retrieved_users = json.loads(file.read())
-        os.remove(users_filename)
-
-        # ensure that an export can be imported
-        self._import_users_from_file(retrieved_users)
-
-        def find_by_username(username):
-            matches = [u for u in retrieved_users
-                       if u['username'] == username]
-            if not matches:
-                self.fail("Couldn't find user with username {}".format(username))
-            else:
-                matches[0].pop('id')  # this key not required for import
-                return matches[0]
-
-        self.assertEqual(find_by_username('imported_user1'), user1)
-        self.assertEqual(find_by_username('imported_user2'), user2)
-
-    def _import_users_from_file(self, user_list):
-        json_file_content = json.dumps(user_list)
-        f = NamedTemporaryFile(delete=False)
-        try:
-            f.write(json_file_content.encode())
-            f.flush()
-
-            args = self.parser.parse_args([
-                'users', 'import', f.name
-            ])
-            cli.users_import(args)
-        finally:
-            os.remove(f.name)
-
-    def _export_users_to_file(self):
-        f = NamedTemporaryFile(delete=False)
-        args = self.parser.parse_args([
-            'users', 'export', f.name
-        ])
-        cli.users_export(args)
-        return f.name
-
-    def _does_user_belong_to_role(self, email, rolename):
-        user = self.appbuilder.sm.find_user(email=email)
-        role = self.appbuilder.sm.find_role(rolename)
-        if user and role:
-            return role in user.roles
-
-        return False
-
-    def test_cli_add_user_role(self):
-        args = self.parser.parse_args([
-            'users', 'create', '--username', 'test4', '--lastname', 'doe',
-            '--firstname', 'jon',
-            '--email', self.TEST_USER1_EMAIL, '--role', 'Viewer', '--use_random_password'
-        ])
-        cli.users_create(args)
-
-        self.assertFalse(
-            self._does_user_belong_to_role(email=self.TEST_USER1_EMAIL,
-                                           rolename='Op'),
-            "User should not yet be a member of role 'Op'"
-        )
-
-        args = self.parser.parse_args([
-            'users', 'add_role', '--username', 'test4', '--role', 'Op'
-        ])
-        cli.users_manage_role(args, remove=False)
-
-        self.assertTrue(
-            self._does_user_belong_to_role(email=self.TEST_USER1_EMAIL,
-                                           rolename='Op'),
-            "User should have been added to role 'Op'"
-        )
-
-    def test_cli_remove_user_role(self):
-        args = self.parser.parse_args([
-            'users', 'create', '--username', 'test4', '--lastname', 'doe',
-            '--firstname', 'jon',
-            '--email', self.TEST_USER1_EMAIL, '--role', 'Viewer', '--use_random_password'
-        ])
-        cli.users_create(args)
-
-        self.assertTrue(
-            self._does_user_belong_to_role(email=self.TEST_USER1_EMAIL,
-                                           rolename='Viewer'),
-            "User should have been created with role 'Viewer'"
-        )
-
-        args = self.parser.parse_args([
-            'users', 'remove_role', '--username', 'test4', '--role', 'Viewer'
-        ])
-        cli.users_manage_role(args, remove=True)
-
-        self.assertFalse(
-            self._does_user_belong_to_role(email=self.TEST_USER1_EMAIL,
-                                           rolename='Viewer'),
-            "User should have been removed from role 'Viewer'"
-        )
-
-    @mock.patch("airflow.bin.cli.DagBag")
-    def test_cli_sync_perm(self, dagbag_mock):
-        self.expect_dagbag_contains([
-            DAG('has_access_control',
-                access_control={
-                    'Public': {'can_dag_read'}
-                }),
-            DAG('no_access_control')
-        ], dagbag_mock)
-        self.appbuilder.sm = mock.Mock()
-
-        args = self.parser.parse_args([
-            'sync_perm'
-        ])
-        cli.sync_perm(args)
-
-        assert self.appbuilder.sm.sync_roles.call_count == 1
-
-        self.assertEqual(2,
-                         len(self.appbuilder.sm.sync_perm_for_dag.mock_calls))
-        self.appbuilder.sm.sync_perm_for_dag.assert_any_call(
-            'has_access_control',
-            {'Public': {'can_dag_read'}}
-        )
-        self.appbuilder.sm.sync_perm_for_dag.assert_any_call(
-            'no_access_control',
-            None,
-        )
-
-    def expect_dagbag_contains(self, dags, dagbag_mock):
-        dagbag = mock.Mock()
-        dagbag.dags = {dag.dag_id: dag for dag in dags}
-        dagbag_mock.return_value = dagbag
-
-    def test_cli_create_roles(self):
-        self.assertIsNone(self.appbuilder.sm.find_role('FakeTeamA'))
-        self.assertIsNone(self.appbuilder.sm.find_role('FakeTeamB'))
-
-        args = self.parser.parse_args([
-            'roles', 'create', 'FakeTeamA', 'FakeTeamB'
-        ])
-        cli.roles_create(args)
-
-        self.assertIsNotNone(self.appbuilder.sm.find_role('FakeTeamA'))
-        self.assertIsNotNone(self.appbuilder.sm.find_role('FakeTeamB'))
-
-    def test_cli_create_roles_is_reentrant(self):
-        self.assertIsNone(self.appbuilder.sm.find_role('FakeTeamA'))
-        self.assertIsNone(self.appbuilder.sm.find_role('FakeTeamB'))
-
-        args = self.parser.parse_args([
-            'roles', 'create', 'FakeTeamA', 'FakeTeamB'
-        ])
-
-        cli.roles_create(args)
-
-        self.assertIsNotNone(self.appbuilder.sm.find_role('FakeTeamA'))
-        self.assertIsNotNone(self.appbuilder.sm.find_role('FakeTeamB'))
-
-    def test_cli_list_roles(self):
-        self.appbuilder.sm.add_role('FakeTeamA')
-        self.appbuilder.sm.add_role('FakeTeamB')
-
-        with mock.patch('sys.stdout', new_callable=io.StringIO) as mock_stdout:
-            cli.roles_list(self.parser.parse_args(['roles', 'list']))
-            stdout = mock_stdout.getvalue()
-
-        self.assertIn('FakeTeamA', stdout)
-        self.assertIn('FakeTeamB', stdout)
-
-    def test_cli_list_roles_with_args(self):
-        cli.roles_list(self.parser.parse_args(['roles', 'list',
-                                               '--output', 'tsv']))
 
     @mock.patch("airflow.bin.cli.db.initdb")
     def test_cli_initdb(self, initdb_mock):
